refactor(web): extract shared PuzzleDefinitionFields from suggest-edit
